+++ testData/org/elixir_lang/heex/reference/embedded_call_find_usages/delegate/page_live.ex
@@ -1,0 +1,3 @@
+defmodule MyAppWeb.PageLive do
+  defdelegate delegated_function(), to: MyAppWeb.PageLive.Helper
+end